+++ main.site.go
@@ -3,8 +3,9 @@ package main
 import (
 	"encoding/xml"
 	"fmt"
-	"github.com/hornbill/goApiLib"
 	"strings"
+
+	apiLib "github.com/hornbill/goApiLib"
 )
 
 // siteInCache -- Function to check if passed-thorugh site name has been cached
@@ -29,6 +30,7 @@ func searchSite(siteName string, espXmlmc *apiLib.XmlmcInstStruct) (bool, int) {
 	boolReturn := false
 	intReturn := 0
 	//-- ESP Query for site
+	espXmlmc.SetParam("application", "com.hornbill.core")
 	espXmlmc.SetParam("entity", "Site")
 	espXmlmc.SetParam("matchScope", "all")
 	espXmlmc.OpenElement("searchFilter")